Fixed updated library conflicts. Moved to Python3.
